Catch up to template change
Shortly after I set up this repo, we merged a change to
the template for these lambdas, replacing safety with
pip-audit. This applies that upstream change to this
repository.

It also includes the output of the initial `make install`
command, so the changes to pipfile.lock may be
greater than _just_ those made upstream.