Rename my_function to predict during initial setup
** Why are these changes being introduced:

The template readme has a checklist to complete for initial setup.

** Relevant ticket(s):

n/a

** How does this address that need:

The big work here is to update the "my_function" name to "predict" in
all the places that I can find it. I think I've caught them all, but am
not 100% sure.

I have other questions about the setup checklist, which I'll raise
during code review.

** Document any side effects to this change:

The tests pass, but attempting to invoke the lambda via the command
line with a curl command is failing because of an extra argument being
passed. When I update lambda_handler to account for it, then the tests
fail. Will address this during code review.