Add context dictionary support to lambda
** Why are these changes being introduced:

* At the moment, running the curl command from the project readme in a
terminal results in an argument error, rather than the stated response.

** Relevant ticket(s):

n/a

** How does this address that need:

* This adds a second dictionary as an argument to the lambda_handler
method in predict.py. The same change is made to the two tests which
actually call this fuction.

This now matches what the s3-bagit lambda is expecting, which I'm using
as a starting point for my work in this repo:

Handler:
https://github.com/MITLibraries/s3-bagit-validator/blob/main/lambdas/validator.py#L29

Test:
https://github.com/MITLibraries/s3-bagit-validator/blob/main/tests/test_lambda_handler.py#L20

** Document any side effects to this change:

I'm not sure whether this is going to bite me later or not - that will
need to wait until this gets deployed to AWS.