Complete initial setup
** Why are these changes being introduced:

* The template repo has a checklist of steps to perform for initial
setup, which were not all completed in the last PR.

** Relevant ticket(s):

n/a

** How does this address that need:

This updates the project readme to remove that checklist, now that all
the steps have been completed (at least initially - some steps, like the
repo settings, are subject to further change).

** Document any side effects to this change:

This also performs some additional cleanup to the readme, including an
interaction diagram from the engineering plan as well as some other
introductory text.