Sets up SAM, adds ping route
** Why are these changes being introduced:

* The lambda needs to handle HTTP requests in order to communicate with
TACOS, and one way to do that is to adopt SAM.

* As part of this adoption, the lambda also needs to have something to
respond with, and our ping-pong pattern is such an easy feature.

** Relevant ticket(s):

n/a

** How does this address that need:

* This adopts the SAM pattern, following the examples I see with the APT
and s3-bagit lambdas. It also cribs the ping-pong response pattern from
those examples.

** Document any side effects to this change:

* Probably, but I'm new enough to lambdas at this point that I don't
know if I recognize side effects clearly. Feedback and pointers are
welcome.