Extend predict event to include features object
** Why are these changes being introduced:

* In order to actually do work, the "predict" action in this lambda will
need to receive a dictionary of features that will be fed into the model
in order to make the prediction.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-182

** How does this address that need:

* This extends the lambda definition to require a "features" argument
when the "predict" action is requested. We add a JSON schema to enforce
the structure of this dictionary.

* There is a new make command to peform a sample prediction, in addition
to the sample ping action.

** Document any side effects to this change:

Things that may be considered side effects here:

* Payload validation is now spread across two schemes - the top level
arguments provided to the lambda (action, challenge_secret, and the
features dictionary) are enforced by the InputPayload. However, the
contents of the features dictionary are validated by the JSON schema.
This feels a bit awkward, but I wasn't able to easily enforce this
using only the InputPayload approach.

* The fact that we are enforcing the structure of the features
dictionary means that adding parameters to our models will require this
lambda to be updated and re-deployed. This feels acceptable for now,
but we should keep an eye on the development experience here if we do
end up changing features frequently.

* The structure of test/conftest.py is starting to feel a little rough
but I'm not sure what other approaches exist for fixture management
in Python.