GitHub Actions Automation
Why these changes are being introduced:
As part of our standard process, application repositories that build
containers to push to AWS ECR get updates from the Terraform repository
that builds the ECR infrastructure. These updates include commands for
the Makefile that make it easier to build and deploy containers to AWS
ECR and a GitHub Actions workflow for automatically updating the
container image in Dev1 for any PRs to the main branch.

How this addresses that need:
* Updates to the Makefile for the dist-dev and publish-dev commands
* A new GitHub Actions workflow for automatic builds to Dev1 on PRs to
the main branch
* A new GHA workflow for automatic builds to Stage-Workloads on merges
to the main branch
* A new GHA workflow for automatic builds to Prod-Workloads on tagged
releases on the main branch
* Update to README to link back to the related infrastructure
repositories

Side effects of this change:
Not a side effect, but a real effect: Once this is merged to main, the
automated build process will kick in for dev/stage/prod.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TCO-173