Implement real prediction logic (w/side effects)
** Why are these changes being introduced:

The lambda needs to actually do work, not just always return true.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-157

** How does this address that need:

Buckle up, this does a lot. Some of it brings side effects, which I'll
try to describe accurately.

1. This copies an already-trained model from the TACOS notebooks into
this repo. The command I used was:

cp ../tacos-notebooks/notebooks/algorithms/models/neural.pkl lambdas/models/neural.pkl

This probably needs to be automated, but baby steps there.

2. A few new dependencies get added (pandas, pandas-stubs, and
scikit-learn). We're also using the core library's support for pickle
loading.

3. The PredictHandler class now has a load_model method, with a hard-
coded file path directly in this repository. It also includes a check
that the model has been fitted, as a defense against the model building
workflow being _so_ manual.

4. The PredictHandler class' handle method uses the model to make a real
prediction, and returns that as part of its JSON response.

5. The Makefile's use of the predict function gets a bit more usable,
now taking a file argument - with some example payloads in the tests/sam
location. This will hopefully help our DX as we keep working with this
feature.

6. Tests and fixtures should cover all the new functionality, although
most of the tests are more at the integration level (testing everything
from the predict.lambda_handler starting point). I'm still working out
the best testing approach for this codebase - some tests are more
targeted.

Please note that I've elected to keep all the test fixtures in conftest,
rather than use the provided example payloads in tests/sam. The fixtures
and examples should probably be kept in sync, but it's not required -
and I particularly don't want drift in the examples to help DX to end up
requiring changes to the test logic over time. The records are the same
only incidentally.

** Document any side effects to this change:

1. This makes the choice to implement the neural network algorithm,
rather than the other candidates. I chose this because it has been
performing better than the others, but k-nearest neighbors is nearly at
this level so it's not a clear-cut choice.

2. The model as-implemented is still in a branch of the TACOS notebooks,
because the feature handling isn't quite uniform yet. We are moving to
using lowercase feature names, and presenting them in alphabetical
order. Because the model is hand-copied, however, the bespoke process is
not as repeatable nor automatable as it should be.

3. Our code quality tooling isn't happy about using scikit-learn, which
does not support type hinting in the way we require. An exception has
been added to pyproject.toml to acknowledge this (and allow the linter
to still work)

4. With the model being placed directly in this repository, we may be
able to skip the check_is_fitten logic on every execution in favor of
a CI check via Github Actions? If we get to loading the model from S3
then a runtime defense is probably the way to go, so we're now in an
intermediate state.

5. Seriously, perspectives on how to write - and organize - targeted
tests for this codebase would be welcome. Rails provides a very
detailed test structure that allows for a variety of testing scopes,
but I'm not sure if that'd be overkill here or not.