Respond to code review feedback
This makes a few changes that have come about while responding to code
review:

1. We change the return from the prediction to be a boolean, rather than a
string. There will be a related change to TACOS to deal with this.

2. We back away entirely from using check_is_fitted during the lambda's
operation. As Graham and I talked about the issues at play, and what the
proper response should be to an unfitted model, we realized that runtime is
not the concern - given that the model is being committed to version control,
the proper defensive point is during the PR, and as part of unit tests.

It turns out that several of our existing unit tests blow up when an unfitted
model is used in the codebase. One of those has been explicitly renamed as a
test against an unfitted model, although two other tests (the tests of the
positive and negative prediction outcomes) will also fail if the model is not
fitted.

Side effect:
The application now has something of an unprotected belly, which can be
exposed if we ever get to the point of placing the model in an s3 bucket
for loading at runtime rather than deploying after code review and unit
testing.

3. The Makefile gets changed because I'd reversed the slash when declaring
the expected content type. It worked locally, but failed when deployed to
AWS.

4. We also remove a deprecated linter from the ignore list in pyproject.toml
upon feedback from the linter.