Remove lastnames feature, switch to knn model
** Why are these changes being introduced:

* We are removing the lastnames feature from consideration when
detecting citations, which means we need to update the lambda to drop
support for that parameter.

* As a result of dropping the lastnames feature, the training process
now indicates than a kNN algorithm will perform better than a neural
network.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-155

** How does this address that need:

This copies the new model from the notebooks repo, replacing the neural
network algorithm.

We also drop support for the lastnames feature, removing it from the
schema as well as all the tests and fixtures.

** Document any side effects to this change:

As mentioned above, this change involves changing which algorithm is
used to predict whether a search string is a citation or not. The new
algorithm performs slightly worse than before (f1 score in the 0.92
range rather than 0.95), but the neural network degrades more strongly,
going down to about 0.90.