Update model after filtering training data
** Why are these changes being introduced:

* We recently implemented a filter inside TACOS, preventing search terms
with fewer than three non-zero feature values from being sent to the
lambda. We anticipate that this will remove about 90% of searches from
consideration by the lambda - which changes the distribution of
characteristics which the model will experience.

Because of this change, we've elected to retrain the models on an
updated set of records (we applied the same filter logic to our existing
set of labelled records).

The impact of that change is that not only did the model change, but
there is a new best-performing algorithm - a neural network now
outperforms the k-NN algorithm. So we swap models as well.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-191

** How does this address that need:

* We replace the knn model with the newly-upated neural network model.
* The lambda code references the new filename.

** Document any side effects to this change:

* Model performance changes very slightly as a result of this change.
Both the prior model and the new model scored a 0.93 on an f1 score,
but there was a slight change beyond that number of significant digits.